+++ libraries/exoplayer_rtsp/src/test/java/androidx/media3/exoplayer/rtsp/RtspPlaybackTest.java
@@ -46,7 +46,9 @@ import com.google.common.collect.ImmutableList;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 import java.util.concurrent.ConcurrentLinkedQueue;
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicReference;
 import javax.net.SocketFactory;
 import org.junit.After;
@@ -55,6 +57,7 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.robolectric.annotation.Config;
+import org.robolectric.shadows.ShadowLooper;
 
 /** Playback testing for RTSP. */
 @Config(sdk = 29)
@@ -103,9 +106,12 @@ public final class RtspPlaybackTest {
         new ResponseProvider(
             clock,
             ImmutableList.of(aacRtpPacketStreamDump, mpeg2tsRtpPacketStreamDump),
-            fakeRtpDataChannel);
+            fakeRtpDataChannel,
+            RtspMessageUtil.DEFAULT_RTSP_TIMEOUT_MS,
+            /* optionsRequestCounter= */ Optional.empty());
     rtspServer = new RtspServer(responseProvider);
-    ExoPlayer player = createExoPlayer(rtspServer.startAndGetPortNumber(), rtpDataChannelFactory);
+    ExoPlayer player =
+        createExoPlayer(clock, rtspServer.startAndGetPortNumber(), rtpDataChannelFactory);
 
     PlaybackOutput playbackOutput = PlaybackOutput.register(player, capturingRenderersFactory);
     player.prepare();
@@ -125,8 +131,13 @@ public final class RtspPlaybackTest {
     rtspServer =
         new RtspServer(
             new ResponseProvider(
-                clock, ImmutableList.of(mpeg2tsRtpPacketStreamDump), fakeRtpDataChannel));
-    ExoPlayer player = createExoPlayer(rtspServer.startAndGetPortNumber(), rtpDataChannelFactory);
+                clock,
+                ImmutableList.of(mpeg2tsRtpPacketStreamDump),
+                fakeRtpDataChannel,
+                RtspMessageUtil.DEFAULT_RTSP_TIMEOUT_MS,
+                /* optionsRequestCounter= */ Optional.empty()));
+    ExoPlayer player =
+        createExoPlayer(clock, rtspServer.startAndGetPortNumber(), rtpDataChannelFactory);
 
     AtomicReference<Throwable> playbackError = new AtomicReference<>();
     player.prepare();
@@ -158,7 +169,7 @@ public final class RtspPlaybackTest {
             new UdpDataSourceRtpDataChannelFactory(DEFAULT_TIMEOUT_MS), rtpTcpDataChannelFactory);
     rtspServer = new RtspServer(responseProviderSupportingOnlyTcp);
     ExoPlayer player =
-        createExoPlayer(rtspServer.startAndGetPortNumber(), forwardingRtpDataChannelFactory);
+        createExoPlayer(clock, rtspServer.startAndGetPortNumber(), forwardingRtpDataChannelFactory);
 
     PlaybackOutput playbackOutput = PlaybackOutput.register(player, capturingRenderersFactory);
     player.prepare();
@@ -183,7 +194,8 @@ public final class RtspPlaybackTest {
             ImmutableList.of(aacRtpPacketStreamDump, mpeg2tsRtpPacketStreamDump),
             fakeUdpRtpDataChannel);
     rtspServer = new RtspServer(responseProvider);
-    ExoPlayer player = createExoPlayer(rtspServer.startAndGetPortNumber(), rtpDataChannelFactory);
+    ExoPlayer player =
+        createExoPlayer(clock, rtspServer.startAndGetPortNumber(), rtpDataChannelFactory);
 
     AtomicReference<PlaybackException> playbackError = new AtomicReference<>();
     player.prepare();
@@ -220,7 +232,7 @@ public final class RtspPlaybackTest {
         new ForwardingRtpDataChannelFactory(rtpDataChannelFactory, rtpDataChannelFactory);
     rtspServer = new RtspServer(responseProviderSupportingOnlyTcp);
     ExoPlayer player =
-        createExoPlayer(rtspServer.startAndGetPortNumber(), forwardingRtpDataChannelFactory);
+        createExoPlayer(clock, rtspServer.startAndGetPortNumber(), forwardingRtpDataChannelFactory);
 
     AtomicReference<PlaybackException> playbackError = new AtomicReference<>();
     player.prepare();
@@ -240,8 +252,39 @@ public final class RtspPlaybackTest {
     assertThat(playbackError.get()).hasCauseThat().hasMessageThat().isEqualTo("SETUP 461");
   }
 
+  @Test
+  public void play_withCustomSessionTimeoutDuration_sendsKeepAliveOptionsRequest()
+      throws Exception {
+    FakeUdpDataSourceRtpDataChannel fakeRtpDataChannel = new FakeUdpDataSourceRtpDataChannel();
+    RtpDataChannel.Factory rtpDataChannelFactory = (trackId) -> fakeRtpDataChannel;
+    FakeClock fakeClock = new FakeClock(/* initialTimeMs= */ 0, true);
+    Optional<AtomicInteger> optionsRequestCounter = Optional.of(new AtomicInteger());
+    ResponseProvider responseProvider =
+        new ResponseProvider(
+            fakeClock,
+            ImmutableList.of(aacRtpPacketStreamDump),
+            fakeRtpDataChannel,
+            /* sessionTimeoutMs= */ 30_000L,
+            optionsRequestCounter);
+    rtspServer = new RtspServer(responseProvider);
+    ExoPlayer player =
+        createExoPlayer(fakeClock, rtspServer.startAndGetPortNumber(), rtpDataChannelFactory);
+    player.prepare();
+    player.play();
+    TestPlayerRunHelper.runUntilPlaybackState(player, Player.STATE_READY);
+    // Reset optionsRequestCounter to count requests made by the keep-alive monitor
+    optionsRequestCounter.get().getAndSet(0);
+
+    fakeClock.advanceTime(/* timeDiffMs= */ 16_000L);
+    ShadowLooper.idleMainLooper();
+
+    assertThat(optionsRequestCounter.get().get()).isEqualTo(1);
+
+    player.release();
+  }
+
   private ExoPlayer createExoPlayer(
-      int serverRtspPortNumber, RtpDataChannel.Factory rtpDataChannelFactory) {
+      Clock clock, int serverRtspPortNumber, RtpDataChannel.Factory rtpDataChannelFactory) {
     ExoPlayer player =
         new ExoPlayer.Builder(applicationContext, capturingRenderersFactory)
             .setClock(clock)
@@ -260,11 +303,14 @@ public final class RtspPlaybackTest {
   private static class ResponseProvider implements RtspServer.ResponseProvider {
 
     protected static final String SESSION_ID = "00000000";
+    private static final String SESSION_TIMEOUT_HEADER_TAG = ";timeout=";
 
     protected final Clock clock;
-    protected final ArrayList<RtpPacketStreamDump> dumpsForSetUpTracks;
+    protected final List<RtpPacketStreamDump> dumpsForSetUpTracks = new ArrayList<>();
     protected final ImmutableList<RtpPacketStreamDump> rtpPacketStreamDumps;
     private final RtspMessageChannel.InterleavedBinaryDataListener binaryDataListener;
+    private final long sessionTimeoutMs;
+    private final Optional<AtomicInteger> optionsRequestCounter;
 
     protected RtpPacketTransmitter packetTransmitter;
 
@@ -275,15 +321,21 @@ public final class RtspPlaybackTest {
      * @param rtpPacketStreamDumps A list of {@link RtpPacketStreamDump}.
      * @param binaryDataListener A {@link RtspMessageChannel.InterleavedBinaryDataListener} to send
      *     RTP data.
+     * @param sessionTimeoutMs Duration RTSP server will keep the session active without receiving
+     *     any requests.
+     * @param optionsRequestCounter for how many RTSP Options requests were sent.
      */
-    public ResponseProvider(
+    ResponseProvider(
         Clock clock,
         List<RtpPacketStreamDump> rtpPacketStreamDumps,
-        RtspMessageChannel.InterleavedBinaryDataListener binaryDataListener) {
+        RtspMessageChannel.InterleavedBinaryDataListener binaryDataListener,
+        long sessionTimeoutMs,
+        Optional<AtomicInteger> optionsRequestCounter) {
       this.clock = clock;
       this.rtpPacketStreamDumps = ImmutableList.copyOf(rtpPacketStreamDumps);
       this.binaryDataListener = binaryDataListener;
-      dumpsForSetUpTracks = new ArrayList<>();
+      this.sessionTimeoutMs = sessionTimeoutMs;
+      this.optionsRequestCounter = optionsRequestCounter;
     }
 
     /** Returns a list of the received SETUP requests' corresponding {@link RtpPacketStreamDump}. */
@@ -295,6 +347,7 @@ public final class RtspPlaybackTest {
 
     @Override
     public RtspResponse getOptionsResponse() {
+      optionsRequestCounter.ifPresent(AtomicInteger::getAndIncrement);
       return new RtspResponse(
           /* status= */ 200,
           new RtspHeaders.Builder()
@@ -316,9 +369,15 @@ public final class RtspPlaybackTest {
           packetTransmitter = new RtpPacketTransmitter(rtpPacketStreamDump, clock);
         }
       }
-
       return new RtspResponse(
-          /* status= */ 200, headers.buildUpon().add(RtspHeaders.SESSION, SESSION_ID).build());
+          /* status= */ 200,
+          headers
+              .buildUpon()
+              .add(
+                  RtspHeaders.SESSION,
+                  // Convert sessionTimeoutMs to seconds
+                  SESSION_ID + SESSION_TIMEOUT_HEADER_TAG + (sessionTimeoutMs / 1000))
+              .build());
     }
 
     @Override
@@ -348,7 +407,12 @@ public final class RtspPlaybackTest {
         Clock clock,
         List<RtpPacketStreamDump> rtpPacketStreamDumps,
         RtspMessageChannel.InterleavedBinaryDataListener binaryDataListener) {
-      super(clock, rtpPacketStreamDumps, binaryDataListener);
+      super(
+          clock,
+          rtpPacketStreamDumps,
+          binaryDataListener,
+          RtspMessageUtil.DEFAULT_RTSP_TIMEOUT_MS,
+          /* optionsRequestCounter= */ Optional.empty());
     }
 
     @Override